TextRecognizer added.
