+++ ReceiptRecognizer/Program.cs
@@ -1,6 +1,5 @@
-﻿using Emgu.CV;
-using ReceiptRecognizer.ObjectRecognizer;
-using System;
+﻿using ReceiptRecognizer.ObjectRecognizer;
+using ReceiptRecognizer.TextRecognizer;
 
 namespace ReceiptRecognizer
 {
@@ -8,7 +7,7 @@ namespace ReceiptRecognizer
     {
         static void Main(string[] args)
         {
-            var recognizer = new ReceiptRecognizer(new EmguLargestAreaRecognizer());
+            var recognizer = new ReceiptRecognizer(new EmguLargestAreaRecognizer(), new TesseractRecognizer());
             recognizer.RecognizeTestImages();
         }
     }
