+++ ReceiptRecognizer/TextRecognizer/ITextRecognizer.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Drawing;
 using System.Text;
 
-namespace ReceiptRecognizer
+namespace ReceiptRecognizer.TextRecognizer
 {
     public interface ITextRecognizer
     {
+        public string GetText(Bitmap image);
     }
 }
